+++ packages/ozone/src/db/schema/index.ts
@@ -16,6 +16,7 @@ import * as recordEventsStats from './record_events_stats'
 import * as recordPushEvent from './record_push_event'
 import * as repoPushEvent from './repo_push_event'
 import * as report from './report'
+import * as reportQueue from './report_queue'
 import * as safelink from './safelink'
 import * as scheduledAction from './scheduled-action'
 import * as setting from './setting'
@@ -26,6 +27,7 @@ import * as moderatorAssignment from './moderator_assignment'
 export type DatabaseSchemaType = modEvent.PartialDB &
   modSubjectStatus.PartialDB &
   report.PartialDB &
+  reportQueue.PartialDB &
   label.PartialDB &
   signingKey.PartialDB &
   repoPushEvent.PartialDB &